compose_validate: Explode validation_error into individual functions.
